ref(project): Move types to utils
