+++ src/components/app/app.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 
 import classNames from 'classnames';
 
@@ -11,24 +10,6 @@ import appStyles from './app.module.css';
 
 import { ingredients, defaultBasket } from '../../utils/data';
 
-export const ingredientsPropTypes = PropTypes.shape({
-  _id: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  type: PropTypes.oneOf(['bun', 'main', 'sauce']),
-  proteins: PropTypes.number,
-  fat: PropTypes.number,
-  carbohydrates: PropTypes.number,
-  calories: PropTypes.number,
-  price: PropTypes.number.isRequired,
-  image: PropTypes.string.isRequired,
-  image_mobile: PropTypes.string,
-  image_large: PropTypes.string,
-  __v: PropTypes.number
-});
-
-export const basketPropTypes = PropTypes.shape({
-  _id: PropTypes.string.isRequired
-});
 
 const App = () => {
   const [state, setState] = React.useState({
