+++ src/utils/types.js
@@ -1,0 +1,20 @@
+import PropTypes from 'prop-types';
+
+export const ingredientsPropTypes = PropTypes.shape({
+    _id: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+    type: PropTypes.oneOf(['bun', 'main', 'sauce']),
+    proteins: PropTypes.number,
+    fat: PropTypes.number,
+    carbohydrates: PropTypes.number,
+    calories: PropTypes.number,
+    price: PropTypes.number.isRequired,
+    image: PropTypes.string.isRequired,
+    image_mobile: PropTypes.string,
+    image_large: PropTypes.string,
+    __v: PropTypes.number
+});
+
+export const basketPropTypes = PropTypes.shape({
+    _id: PropTypes.string.isRequired
+});
